feat(website): improve typings for settings mappers
