+++ website/package.json
@@ -68,7 +68,7 @@
         "react-helmet": "^6.1.0",
         "react-icons": "^4.3.1",
         "react-markdown": "^8.0.7",
-        "react-select": "^4.3.1",
+        "react-select": "^5.10.1",
         "styled-components": "^5.3.10"
     },
     "scripts": {
