+++ website/src/pages/bar/index.tsx
@@ -1,53 +1,67 @@
 import React from 'react'
-import { ResponsiveBarCanvas } from '@nivo/bar'
+import { graphql, useStaticQuery } from 'gatsby'
+import { patternDotsDef, patternLinesDef } from '@nivo/core'
+import { ResponsiveBar, svgDefaultProps } from '@nivo/bar'
 import { ComponentTemplate } from '../../components/components/ComponentTemplate'
 import meta from '../../data/components/bar/meta.yml'
-import { generateHeavyDataSet } from '../../data/components/bar/generator'
-import mapper from '../../data/components/bar/mapper'
+import { generateLightDataSet } from '../../data/components/bar/generator'
+import mapper, { UnmappedBarProps, MappedBarProps } from '../../data/components/bar/mapper'
 import { groups } from '../../data/components/bar/props'
-import { graphql, useStaticQuery } from 'gatsby'
 
-const Tooltip = data => {
-    /* return custom tooltip */
-}
-
-const initialProperties = {
+const initialProperties: UnmappedBarProps = {
     indexBy: 'country',
 
     margin: {
         top: 50,
-        right: 60,
+        right: 130,
         bottom: 50,
         left: 60,
     },
 
-    pixelRatio:
-        typeof window !== 'undefined' && window.devicePixelRatio ? window.devicePixelRatio : 1,
-
-    padding: 0.15,
+    padding: 0.3,
     innerPadding: 0,
     minValue: 'auto',
     maxValue: 'auto',
 
     groupMode: 'stacked',
-    layout: 'horizontal',
+    layout: 'vertical',
     reverse: false,
 
     valueScale: { type: 'linear' },
     indexScale: { type: 'band', round: true },
     valueFormat: { format: '', enabled: false },
 
-    colors: { scheme: 'red_blue' },
+    colors: { scheme: 'nivo' },
     colorBy: 'id',
-    borderWidth: 0,
+    defs: [
+        patternDotsDef('dots', {
+            background: 'inherit',
+            color: '#38bcb2',
+            size: 4,
+            padding: 1,
+            stagger: true,
+        }),
+        patternLinesDef('lines', {
+            background: 'inherit',
+            color: '#eed312',
+            rotation: -45,
+            lineWidth: 6,
+            spacing: 10,
+        }),
+    ],
+    fill: [
+        { match: { id: 'fries' }, id: 'dots' },
+        { match: { id: 'sandwich' }, id: 'lines' },
+    ],
     borderRadius: 0,
+    borderWidth: 0,
     borderColor: {
         from: 'color',
         modifiers: [['darker', 1.6]],
     },
 
     axisTop: {
-        enable: true,
+        enable: false,
         tickSize: 5,
         tickPadding: 5,
         tickRotation: 0,
@@ -71,7 +85,7 @@ const initialProperties = {
         tickRotation: 0,
         legend: 'country',
         legendPosition: 'middle',
-        legendOffset: 36,
+        legendOffset: 32,
         truncateTickAt: 0,
     },
     axisLeft: {
@@ -85,8 +99,8 @@ const initialProperties = {
         truncateTickAt: 0,
     },
 
-    enableGridX: true,
-    enableGridY: false,
+    enableGridX: false,
+    enableGridY: true,
 
     enableLabel: true,
     enableTotals: false,
@@ -100,21 +114,54 @@ const initialProperties = {
     labelPosition: 'middle',
     labelOffset: 0,
 
+    legends: [
+        {
+            dataFrom: 'keys',
+            anchor: 'bottom-right',
+            direction: 'column',
+            justify: false,
+            translateX: 120,
+            translateY: 0,
+            itemsSpacing: 2,
+            itemWidth: 100,
+            itemHeight: 20,
+            itemDirection: 'left-to-right',
+            itemOpacity: 0.85,
+            symbolSize: 20,
+            onClick: data => {
+                alert(JSON.stringify(data, null, '    '))
+            },
+            effects: [
+                {
+                    on: 'hover',
+                    style: {
+                        itemOpacity: 1,
+                    },
+                },
+            ],
+        },
+    ],
+
     isInteractive: true,
     'custom tooltip example': false,
-    tooltip: null,
 
-    legends: [],
+    animate: true,
+    motionConfig: 'default',
+
+    role: 'application',
+    isFocusable: false,
+    ariaLabel: 'Nivo bar chart demo',
+    barAriaLabel: data => `${data.id}: ${data.formattedValue} in country: ${data.indexValue}`,
 }
 
-const BarCanvas = () => {
+const Bar = () => {
     const {
         image: {
             childImageSharp: { gatsbyImageData: image },
         },
     } = useStaticQuery(graphql`
         query {
-            image: file(absolutePath: { glob: "**/src/assets/captures/bar-canvas.png" }) {
+            image: file(absolutePath: { glob: "**/src/assets/captures/bar.png" }) {
                 childImageSharp {
                     gatsbyImageData(layout: FIXED, width: 700, quality: 100)
                 }
@@ -123,27 +170,28 @@ const BarCanvas = () => {
     `)
 
     return (
-        <ComponentTemplate
-            name="BarCanvas"
-            meta={meta.BarCanvas}
+        <ComponentTemplate<UnmappedBarProps, MappedBarProps, any>
+            name="Bar"
+            meta={meta.Bar}
             icon="bar"
             flavors={meta.flavors}
-            currentFlavor="canvas"
+            currentFlavor="svg"
             properties={groups}
             initialProperties={initialProperties}
+            defaultProperties={svgDefaultProps}
             propertiesMapper={mapper}
             codePropertiesMapper={(properties, data) => ({
                 keys: data.keys,
                 ...properties,
-                tooltip: properties.tooltip ? Tooltip : undefined,
+                tooltip: properties.tooltip ? 'CustomTooltip' : undefined,
             })}
-            generateData={generateHeavyDataSet}
+            generateData={generateLightDataSet}
             getTabData={data => data.data}
             image={image}
         >
             {(properties, data, theme, logAction) => {
                 return (
-                    <ResponsiveBarCanvas
+                    <ResponsiveBar
                         data={data.data}
                         keys={data.keys}
                         {...properties}
@@ -163,4 +211,4 @@ const BarCanvas = () => {
     )
 }
 
-export default BarCanvas
+export default Bar
