+++ website/src/pages/chord/index.tsx
@@ -3,13 +3,13 @@ import { generateChordData } from '@nivo/generators'
 import { ResponsiveChord, svgDefaultProps } from '@nivo/chord'
 import { ComponentTemplate } from '../../components/components/ComponentTemplate'
 import meta from '../../data/components/chord/meta.yml'
-import mapper from '../../data/components/chord/mapper'
+import mapper, { UnmappedChordProps, MappedChordProps } from '../../data/components/chord/mapper'
 import { groups } from '../../data/components/chord/props'
 import { graphql, useStaticQuery } from 'gatsby'
 
 const MATRIX_SIZE = 5
 
-const initialProperties = {
+const initialProperties: UnmappedChordProps = {
     margin: {
         top: 60,
         right: 60,
@@ -105,7 +105,7 @@ const Chord = () => {
     `)
 
     return (
-        <ComponentTemplate
+        <ComponentTemplate<UnmappedChordProps, MappedChordProps, any>
             name="Chord"
             meta={meta.Chord}
             icon="chord"
